% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_commute.R
\name{clean_eqn}
\alias{clean_eqn}
\title{Process Generated Equations for Readability}
\usage{
clean_eqn(eqn)
}
\arguments{
\item{eqn}{A character vector whose elements are terms in an algebraic
equation of the form \code{ax + by +cz...}, where \code{a}, \code{b}, and \code{c} are numeric
coefficients for unknown variables.}
}
\value{
a length-one character vector that drops zero-coefficient terms and
separates terms with the correct \code{+} and \code{-} signs. If all terms are zeroed,
the empty string.
}
\description{
Process Generated Equations for Readability
}
