% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation2latex.R
\name{equation2latex}
\alias{equation2latex}
\title{Transform an R Expression into LaTeX Code}
\usage{
equation2latex(eqn, env, emBiggen = TRUE)
}
\arguments{
\item{eqn}{An R expression, optionally containing matrices.}

\item{env}{Environment in which to evaluate the expression.
Defaults to the caller environment of \code{equation2latex}.}

\item{emBiggen}{Logical determining whether to apply the
LaTeX Bigg modifier to parentheses in the output, which looks
better with matrices. "Embiggens" is a perfectly cromulent word.
Defaults to TRUE.}
}
\value{
LaTeX code for rendering the expression, as a length-one
character vector.
}
\description{
This function converts unquoted R expressions into LaTeX
code. It uses \verb{extract values} to substitute for each symbol
in the expression its value in \code{env}. Then it deparses the
expression and inserts LaTeX code in the appropriate places,
including handling matrices and vectors.
}
\examples{

A <- matrix(c(1, 2, 3, 4), nrow = 2)
B <- A + diag(x = 3, nrow = 2)
C <- matrix(c(7, 8, 9, 10), nrow = 2)
D <- matrix(c(1, 7, 9,-9), nrow = 2)
k <- 7

print_eqn(equation2latex(A))
print_eqn(equation2latex(k *A + (B + C) \%*\% D))


}
