% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation2latex.R
\name{extract_values}
\alias{extract_values}
\title{Extract and Modify Values in an Expression}
\usage{
extract_values(exp, env, pred = function(val) TRUE, fun = identity, ...)
}
\arguments{
\item{exp}{A quoted expression.}

\item{env}{An environment in which to evaluate values in the expression.
Defaults to the caller environment.}

\item{pred}{A predicate function that returns a length-one logical when
applied to evaluated values within the expression. Defaults to \code{TRUE}.}

\item{fun}{An optional function to apply to each evaluated value.
\code{extract_values} will return whatever this function returns for each value
successfully evaluated in \code{env}. Defaults to \code{identity} - which leaves the
value unchanged.}

\item{...}{Additional arguments to \code{fun}.}
}
\value{
A named vector or list of the successfully evaluated values.
If none were successfully evaluated, an empty list.
}
\description{
This function extracts symbols from an expression and
evaluates them in a manner the user specifies. First, each symbol
is evaluated in the supplied \code{envir}. If it is defined in that environment
and is not bound to a function, \code{extract_values} then applies \code{pred} to
the value. If \code{pred} returns \code{TRUE}, the value is then transformed by
\code{fun}.
}
\examples{
 envir <- new.env()
 envir$A <- square(1:9)
 B <- square(4, 5, 6, 7)
 k <- 2
 extract_values(quote(A + B), env = envir, pred = function(x) nrow(x) >2, fun = mean)
 extract_values(quote(k * A), pred = is.matrix, fun = function(x) x \%^\% 3)

}
