% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_inverse.R
\name{gen_inverse}
\alias{gen_inverse}
\title{Solve Systems Using the Generalized Inverse}
\usage{
gen_inverse(A, b)
}
\arguments{
\item{A}{An n x m matrix where n>m}

\item{b}{An n x 1 vector for which to solve Ax=b}
}
\value{
The m x 1 vector x that solve the system, if it exists; otherwise, an
error
}
\description{
This function uses the pseudoinverse method to solve systems of
equations given in non-square matrices.
}
\examples{
gen_inverse(matrix(1:6, nrow = 3), 1:3)
}
