% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_abline2.R
\name{geom_abline2}
\alias{geom_abline2}
\title{Process Parameters for both Vertical and Non-Vertical Lines}
\usage{
geom_abline2(slope, intercept, color)
}
\arguments{
\item{slope}{Argument to geom_abline. If infinite, passed to geom_vline instead.}

\item{intercept}{Argument to geom_abline. Treated as geom_vline's x-intercept argument if slope is infinite.}

\item{color}{Argument to either function}
}
\value{
a geom_abline or geom_vline object, depending on the arguments used
}
\description{
Process Parameters for both Vertical and Non-Vertical Lines
}
