% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_pows.R
\name{\%^\%}
\alias{\%^\%}
\title{Compute Powers of Matrices}
\usage{
m \%^\% pow
}
\arguments{
\item{m}{An \emph{n} x \emph{n} matrix.}

\item{pow}{A nonzero integer, the exponent to compute.}
}
\value{
m raised to the number passed as pow. If pow=1, the unmodified matrix.
}
\description{
Multiplies a square matrix by itself
}
\examples{
m <- matrix(1:9, nrow =3)
m \%^\% 4
}
