% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_axes.R
\name{make_axes}
\alias{make_axes}
\title{Draw X and Y Axes on a Plot}
\usage{
make_axes(p, x_int = 0, y_int = 0)
}
\arguments{
\item{x}{x-intercept of the y-axis, 0 by default}

\item{y}{y-intercept of the x-axis, 0 by default}
}
\value{
A list of four geom_segment objects, each corresponding to one half-axis, that may be added to a ggplot object.
}
\description{
This function creates vertical and horizontal axes that may be added to a ggplot object. The intercepts for either or each may be specified.
Note the axes will not appear if they fall outside the limits of your plot.
This function is partly based on this StackOverflow post: https://stackoverflow.com/questions/17753101/center-x-and-y-axis-with-ggplot2
}
