% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2latex.R
\name{mat2latex}
\alias{mat2latex}
\title{mat2latex
Converts an R matrix object to a representation in Latex code}
\usage{
mat2latex(m)
}
\arguments{
\item{m}{An R matrix object of any dimension}
}
\value{
Latex code for printing the matrix
}
\description{
mat2latex
Converts an R matrix object to a representation in Latex code
}
