% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_pows.R
\name{mat_pows}
\alias{mat_pows}
\title{Compute Powers of a Matrix}
\usage{
mat_pows(m, pows)
}
\arguments{
\item{m}{An \emph{n} x \emph{n} numeric matrix}

\item{pows}{Numeric vector of distinct powers to raise m to. Must be positive natural numbers.}
}
\value{
A list of equal length to pows, each entry of which corresponds to that power of m.
}
\description{
This function takes a square matrix and exponentiates it.
}
\examples{
m1 <- diag(x=3, nrow =4)
m2 <- matrix(c(1, 5, 5, 25), nrow =2) /26
mat_pows(m1, 1:5)
#Projection matrices are idempotent:
mat_pows(m2, c(1, 5, 1000))
}
