% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decomposition.R
\name{plot_decomposition}
\alias{plot_decomposition}
\title{Graphically Display the Orthogonal Decomposition of a Two-Dimensional Matrix}
\usage{
plot_decomposition(
  m,
  color_par = "blue",
  color_perp = "red",
  color_text = "red4",
  output = "plot",
  fix_coords = FALSE
)
}
\arguments{
\item{m}{A 2 x 2 numeric matrix of full rank, one whose columns provide a
valid basis for the two-dimensional plane. If \code{m} has different dimensions,
a different type, or does not have full rank, an error is thrown.}

\item{color_par}{Color for the second vector's parallel component.}

\item{color_perp}{Color for the second vector's orthogonal component.}

\item{color_text}{Color for the text of the equations.}

\item{output}{Output mode. If "plot", a single \code{gridExtra} multiplot featuring
the plots side by side. If "list", a list containing the plots. Defaults to
"plot."}

\item{fix_coords}{Logical determining whether to draw the plots with fixed
aspect ratio. Defaults to FALSE.}
}
\value{
If \code{output} is "plot", a multiplot featuring the six decomposition plots;
otherwise a list containing them.
}
\description{
This function produces plots illustrating the transformation of
a pair of linearly independent two-dimensional vectors into an orthogonal
basis (the \code{Q} of the \code{QR} decomposition). The algorithm, which can be found
in any linear algebra textbook, scales the first vector to unit length, then
extracts the component of the second vector orthogonal to the first and
normalizes it. The plots are based on those found on pp. 218-219 of Otto
Bretscher's \emph{Linear Algebra with Applications} (5th ed.).
}
\examples{
plot_decomposition(square(1, 4, 7, 2))
# With custom colors
plot_decomposition(square(1:4), color_par = "green", color_perp = "orange")
# Already orthogonal vectors: only scaling necessary
plot_decomposition(square(30, 20, 2, -3))
}
