% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{Plot the Lines of the Matrix of a System of Equations}
\usage{
plot_lines(m, b = rep(0, nrow(m)), colors = grDevices::rainbow(n = nrow(m)))
}
\arguments{
\item{m}{An \emph{n} x 2 matrix representing a system of equations in n dimensions}

\item{b}{An \emph{n} x 1 matrix or atomic vector representing the solutions to the equations of \emph{m}. If omitted, a vector of n 0s isused, as in a homoegeneous system of equations}

\item{colors}{Optional vector of colors for the lines, one for each equation in m, reading downward. If not provided, colors are chosen automatically. If shorter than b,
it is recycled per R's usual rules. If longer,extra colors are ignored.}
}
\value{
A ggplot object depicting the lines of the equations.
}
\description{
This function plots the system of equations represented by a matrix. For each row, it plots a line representing every possible
two-dimensional vector that would satisfy the equation.
}
\examples{
m1 <- matrix(c(1,3, 2, 4), nrow =2)
m2 <- matrix(c(1, 3, 2, 6), nrow =2)
# Consistent system
plot_lines(m = m1, b = c(2, 0))
# Inconsistent system
plot_lines(m = m2, b = c(5, 7))
# Linearly dependent system - only one line appears because there are infinitely many solutions
plot_lines(m = m2, b = c(1, 2))
#Plot many equations at once
m3 <- matrix(sample(-10:10, 10), nrow = 5)
plot_lines(m = m3, b = -2:2)
}
