% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trans.R
\name{plot_transform}
\alias{plot_transform}
\title{Plot a Matrix Before and After a Linear Transformation}
\usage{
plot_transform(m, trans, before = "blue", after = "red")
}
\arguments{
\item{m}{A matrix of any dimension for which multiplication by trans (or, if trans is a list, its composition) is defined}

\item{trans}{A 2 x m matrix representing the transformation to apply to m, for which multiplicaiton by m is defined.. Alternately, a list of matrices for which multiplication is defined in the conventional
order. If such a list is passed, the leftmost matrix (the final transformation) must have two dimensions.}

\item{before}{Color to use plotting the vectors of m. May be hexadecimal.}

\item{after}{Color to use when potting the vectors of the transformed matrix. May be hexadecimal.}
}
\value{
A side-by-side plot, the first panel depicting the original matrix, the second the matrix after its transformation. If m is not two-dimensional, only
a plot of the image.
}
\description{
Plot a Matrix Before and After a Linear Transformation
}
\examples{
#45-degree rotation counterclockwise
plot_transform(m = matrix(c(3, 1, 1, 2),
trans = matrix(c(1/sqrt(2), 1/sqrt(2), -1/sqrt(2), 1/sqrt(2))))
}
