% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2latex.R
\name{poly2latex}
\alias{poly2latex}
\title{Convert a Numeric Vector to LaTeX code for a Polynomial Expression}
\usage{
poly2latex(
  v,
  sym = "t",
  dir = "forward",
  lhs = paste0("f(", sym, ") = "),
  sink = FALSE
)
}
\arguments{
\item{v}{A numeric vector of arbitrary length}

\item{sym}{A character to represent the unknown in the polynomial expression.
Defaults to "t"}

\item{dir}{Direction in which the terms of the polynomial represented by the
vector increase. For "forward", the default, the constant term is the leftmost,
for "backward" the rightmost}

\item{lhs}{Value to use to represent the left-hand side of the equation.
Defaults to \verb{f(t) = } Set to "" or NULL to obtain only the right-hand side.}

\item{sink}{Logical determining whether to return the output conventionally or
print it directly to console. Defaults to FALSE.}
}
\value{

}
\description{
This function transforms a numeric vector representing a
polynomial in polynomial space into LaTeX code to render that polynomial. Such
vectors represent linear combinations of coefficient and unknown terms, which
form the basis of the space. This function is intended for use in Rmarkdown
documents with PDF output, where it can convert vectors into formatted
equations automatically.
}
\examples{
poly2latex(c(1:5))
#Constant function
poly2latex(-7)
poly2latex(round(rnorm(5), digits = 2)
}
