% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_eqn.R
\name{print_eqn}
\alias{print_eqn}
\title{Print an Equation for LaTeX Rendering}
\usage{
print_eqn(eqn, delim = "$$", file = "")
}
\arguments{
\item{eqn}{A character vector representing LaTeX code that may be interpreted
by a math environment.}

\item{delim}{Delimiter for the printed code. Defaults to "$$", the TeX standard,
which should suit most uses.}

\item{file}{Passed to \code{cat}'s file argument, with \code{append = FALSE}. Defaults
to "" (i.e., print directly to standard output).}
}
\value{
The object printed to console, or appended to a file if \code{file} is
set.
}
\description{
This simple function takes a a character vector representing an
equation, matrix, or other text interpretable by a LaTeX math environment and
outputs it to console. It is used by \code{mat2latex} and \code{poly2latex} to produce
LaTeX-readable code from R objects.
This function builds on \href{https://stackoverflow.com/questions/30879083/make-r-markdown-code-blocks-into-math-mode}{this} StackOverflow post.
}
\examples{
# A polynomial transformation - note the escape backslashes!
# Try this in an rmarkdown chunk with `results = asis`
eqn <- c("\\\\begin{aligned}", "& T(f) = f(3) + f''\\\\\\\\",
"& = a + 3b + 9c +b + 2ct", "\\\\end{aligned}")
print_eqn(eqn)
}
