% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_ticks.R
\name{size_ticks}
\alias{size_ticks}
\title{Compute Spans for Ticks on custom Axes}
\usage{
size_ticks(breaks, span = 0.0125)
}
\arguments{
\item{breaks}{Vector of breaks for a ggplot object}

\item{span}{Proportion of the range of the breaks the ticks should span in each direction}
}
\value{
Value representing the distance left or right the tick should span from the drawn acis
}
\description{
Compute Spans for Ticks on custom Axes
}
