% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Create a Square Matrix}
\usage{
square(..., byrow = FALSE)
}
\arguments{
\item{...}{Elements of the matrix as vectors whose total number of elements is
a perfect square. Elements of differing types are automatically coerced using
\code{c}.}

\item{byrow}{Argument passed along to \code{matrix}. Defaults to FALSE.}
}
\value{
An n x n matrix, where n is equal to the length of \code{...} after
concatenation.
}
\description{
This simple convenience function creates a square matrix from
elements passed via @param ... . It will throw an error if the number of
elements passed to @param ... is not a perfect square.
}
\examples{
square(1, 5, 6, 7, 8, 9, 1, 2, -1)
# Also works with vectors of differing length
square(-10:0, 5, 2, 3, 0, 9)
}
