% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mat.R
\name{tidy_mat}
\alias{tidy_mat}
\title{Convert a 2 x m Matrix to an m x 2 Tibble for Plotting}
\usage{
tidy_mat(m)
}
\arguments{
\item{m}{A 2 x m matrix.}
}
\value{
A \link[tibble:tibble-package]{tibble} organizing the matrix's values in
plottable form, with columns \emph{x} and \emph{y} and one row for each column vector
of the origianl matrix.
}
\description{
Convert a 2 x m Matrix to an m x 2 Tibble for Plotting
}
\examples{
tidy_mat(matrix(rnorm(n = 20, mean = 5, sd = 3), nrow = 2))
}
