% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_helpers.R
\name{vec_helpers}
\alias{vec_helpers}
\alias{l1}
\alias{l2}
\alias{normalize}
\alias{angle}
\alias{validate_vec}
\alias{dot}
\title{Vector Helper Functions}
\usage{
l1(v)

l2(v)

normalize(v)

dot(v1, v2)

angle(v1, v2)
}
\arguments{
\item{v, v1, v2}{Numeric vectors or 1 x \code{n} matrices. For the binary functions,
\code{v1} and \code{v2} must have the same length (if vectors) or number of rows (if
matrices).}
}
\value{
\code{l1} returns the \code{l1} norm of the vector.

\code{l2} returns the \code{l2} norm of the vector.

\code{normalize} returns the vector normalized to unit length, or
an error if passed the zero vector.

\code{dot} returns the dot product of two vectors.

\code{angle} returns the angle between the two vectors in radians.
}
\description{
These simple functions perform common computations on vectors:
\code{l1} and \code{l2} norms, unit length normalization, dot products, and angles.
}
\examples{
dot(c(2, -1), c(1, 2))
angle(c(1, 0), c(0, 1))
normalize(rep(1, 16))
l1(-2:2)
l2(-2:2)
}
