% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_commute.R
\name{check_commute}
\alias{check_commute}
\title{Symbolically Represent Matrix Commutation}
\usage{
check_commute(
  m,
  syms = matrix(rep_len(letters, length.out = prod(dim(m))), nrow = nrow(m), byrow =
    TRUE),
  latex = FALSE,
  sink = TRUE
)
}
\arguments{
\item{m}{An n x n matrix of dimension greater than 1 whose commutativity will
be tested.}

\item{syms}{Matrix of symbols to use to represent the commuting matrices.
Must have the same dimensions as @param m. Defaults to the first n^2 letters
of the alphabet, entered rowwise, as is conventional. If @param m has
dimension 6 or greater, letters are recycled as needed.}

\item{latex}{is false.}

\item{sink}{Argument passed on to \code{mat2latex}. Defaults to FALSE, ignored if}
}
\value{

}
\description{
This function generates symbolic matrices representing those
that might commute with a particular matrix M (i.e., \code{AM = MA}). This only
applies to square matrices; passing a nonsquare matrix will cause an error.
This function returns a matrix of the input matrix's dimension, with each
element consisting of an equation, where the left side is that element's
value multiplying on the left and the right its value multiplying on the
right. All matrices whose elements satisfy all the equations commute.

All n x n matrices trivially commute with `the corresponding identity and
zero matrix, as well as the inverse, if it exists.
}
\examples{
check_commute(matrix(1:4, nrow = 2))
}
