% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_trans.R
\name{compose_trans}
\alias{compose_trans}
\title{Compose a Series of Matrix Transformations}
\usage{
compose_trans(mats)
}
\arguments{
\item{mats}{A list of matrices that can be multiplied in the conventional order, i.e., A(B(C)). The range of each matrix must equal the domain of the matrix to its right,
per the conventional definition of matrix multiplication. Vectors may also be included if represented as n x 1 matrices, assuming they obey the above constraint.}
}
\value{
The composition of the transformations (i.e., the product of multiplying the matrices in order)
}
\description{
Compose a Series of Matrix Transformations
}
\examples{
a <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3)
b <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2)
compose(list(b,a))
}
