% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_abline2.R
\name{geom_abline2}
\alias{geom_abline2}
\title{Process Parameters for both Vertical and Non-Vertical Lines}
\usage{
geom_abline2(slope, intercept, color)
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Process Parameters for both Vertical and Non-Vertical Lines
}
