% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2latex.R
\name{mat2latex}
\alias{mat2latex}
\title{Convert an R Matrix Object to LaTeX Code}
\usage{
mat2latex(m)
}
\arguments{
\item{m}{An R matrix object of any dimension}
}
\value{
Latex code for printing the matrix. To render the code in an RMarkdown document, call the funciton in a chunk with  the \code{results = "asis"} option.s
}
\description{
Convert an R Matrix Object to LaTeX Code
}
\examples{
#Large matrices are not difficult to process.
mat2latex(as.matrix(mtcars))
}
