% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{plot_lines}
\usage{
plot_lines(m, b, colors = grDevices::rainbow(n = nrow(m)))
}
\arguments{
\item{m}{An n x 2 matrix representing a system of equations in n dimensions}

\item{b}{An n x 1 matrix or atomic vector representing the solutions to the equations of m}

\item{colors}{Optional vector of n distinct colors, one for each equation in m, reading downward. If not provided, colors are chosen automatically}
}
\value{

}
\description{
plot_lines
}
