% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mat.R
\name{plot_mat}
\alias{plot_mat}
\title{Title}
\usage{
plot_mat(m, color = "blue")
}
\arguments{
\item{m}{A two-row matrix with any number of columns. Rows are assumed to represent dimensions}

\item{color}{Color to use when drawing the matrix vectors. may be hexadecimal.}
}
\value{
A ggplot object plotting each vector encoded in the provided matrix
}
\description{
Title
}
\examples{
plot_mat(rbind(-10:0, 0:10))
}
