% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trans.R
\name{plot_transform}
\alias{plot_transform}
\title{Plot a Matrix Before and After a Linear Transformation}
\usage{
plot_transform(m, trans, before = "blue", after = "red")
}
\arguments{
\item{m}{A 2 x m matrix}

\item{trans}{A 2 x 2 matrix representing the transformation to apply to m}

\item{before}{Color to use plotting the vectors of m. May be hexadecimal.}

\item{after}{Color to use when potting the vectors of the transformed matrix. May be hexadecimal.}
}
\value{
A side-by-side plot, the first panel depicting the original matrix, the second the matrix after its transformation.
}
\description{
Plot a Matrix Before and After a Linear Transformation
}
\examples{
#45-degree rotation counterclockwise
plot_transform(m = matrix(c(3, 1, 1, 2),
trans = matrix(c(1/sqrt(2), 1/sqrt(2), -1/sqrt(2), 1/sqrt(2))))
}
